+++ src/constants.py
@@ -7,11 +7,11 @@ def get_root_dir() -> str:
 
     return os.path.dirname(sys.argv[0])
 
-def get_temp_images_dir() -> str:
+def get_temp_images_dir(create: bool = True) -> str:
     """Get the directory of temp_images"""
 
     temp_images_dir = os.path.join(get_root_dir(), "temp_images")
-    if not os.path.exists(temp_images_dir):
+    if not os.path.exists(temp_images_dir) and create:
         os.makedirs(temp_images_dir)
     return temp_images_dir