+++ src/download.py
@@ -1,6 +1,4 @@
 import importlib
-import os
-import pathlib
 import click
 
 from src import constants
@@ -42,8 +40,10 @@ def download(url: str, provider: str | None) -> None:
         downloader(url).download()
     except constants.ProgError as e:
         raise Exception(e)
+    except ModuleNotFoundError:
+        click.echo(f"{provider} is not a valid provider.", err=True)
     except Exception as e:
-        click.echo(f"Error occurred while downloading:\n{e}", err=True)
+        click.echo(f"Error occurred while downloading:\n\t{e}", err=True)
 
 def get_provider() -> str:
     """Gets user selected provider from list of available providers"""